+++ NewHorizons/Patches/CloakPatches.cs
@@ -1,0 +1,31 @@
+using HarmonyLib;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace NewHorizons.Patches
+{
+    public static class CloakPatches
+    {
+        [HarmonyPostfix]
+        [HarmonyPatch(typeof(CloakFieldController), nameof(CloakFieldController.isPlayerInsideCloak), MethodType.Getter)]
+        public static void CloakFieldController_isPlayerInsideCloak(CloakFieldController __instance, ref bool __result)
+        {
+            __result = __result || Components.CloakSectorController.isPlayerInside;
+        }
+
+        [HarmonyPostfix]
+        [HarmonyPatch(typeof(CloakFieldController), nameof(CloakFieldController.isProbeInsideCloak), MethodType.Getter)]
+        public static void CloakFieldController_isProbeInsideCloak(CloakFieldController __instance, ref bool __result)
+        {
+            __result = __result || Components.CloakSectorController.isProbeInside;
+        }
+
+        [HarmonyPostfix]
+        [HarmonyPatch(typeof(CloakFieldController), nameof(CloakFieldController.isShipInsideCloak), MethodType.Getter)]
+        public static void CloakFieldController_isShipInsideCloak(CloakFieldController __instance, ref bool __result)
+        {
+            __result = __result || Components.CloakSectorController.isShipInside;
+        }
+    }
+}
